AGM: fix pointer casting issue
32-bit memory pointer was casted to 64-bit memory. It leads
memory corruption. Change variable and argument type to avoid it.

+++ ipc/HwBinders/agm_ipc_client/src/agm_client_wrapper.cpp
@@ -826,7 +826,7 @@ int agm_session_set_non_tunnel_mode_config(uint64_t handle,
     return -EINVAL;
 }
 
-int agm_session_write_with_metadata(uint64_t handle, struct agm_buff *buf, uint32_t *consumed_size)
+int agm_session_write_with_metadata(uint64_t handle, struct agm_buff *buf, size_t *consumed_size)
 {
     ALOGV("%s called with handle = %x \n", __func__, handle);
     int32_t ret = -EINVAL;
